* origin/master: (65 commits)
  BUG: refs #0458. Changed check to search for first existing parent dir.
  ENH: refs #446. Server side quota validation for uploading new revisions
  ENH: refs #446. Client side validation for upload quota on revision upload
  BUG: refs #0458. Added is_writable check to prevent breaking json parsing.
  ENH: refs #0446. Refactor upload client code
  BUG: refs #0456. Fixes issue where the metadata flowed out of its div.
  ENH: refs #442. Resign applet jar with new certificate. Updated related docs.
  ENH: refs #0448. Add a test for the web api info reporting methods
  ENH: refs #448. Add web api info methods to report enables modules and methods
  STYLE: refs #0445. Userpage Actions links should have mouseover highlighting
  ENH: refs #0445. Show a user's community invitations on their user page
  ENH: refs #445. Private user should be allowed to view their own userpage
  BUG: refs #0445. Alphabet goes all the way to Z
  STYLE: refs #0340. Fix trailing whitespace, and fix stub admin ctrlr test
  ENH: refs #0340. Add stub admin ctrlr test for better coverage reporting
  ENH: refs #340 #444. Add a test for admin setting fuctionality
  ENH: refs #340 #444. Allow admin status to be set on a user by an admin user
  BUG: refs #340 #444. Don't change first/last name display when updating profile
  ENH: refs #0340. Fix test so it won't break other tests
  STYLE: refs #0340. Fix UI look and feel for delete user dialog
  ...

Conflicts:
	core/public/js/layout/main.js
	core/public/js/user/user.userpage.js
	modules/sizequota/public/js/common/common.notify.js

+++ core/models/base/BitstreamModelBase.php
@@ -59,13 +59,14 @@ abstract class BitstreamModelBase extends AppModel
     {
     if(!$bitstream instanceof BitstreamDao)
       {
-      throw new Zend_Exception("Error param.");
+      throw new Zend_Exception('Must pass a bitstream dao');
       }
     $checksum = $bitstream->getChecksum();
     $path = $bitstream->getFullPath();
     $assetstore = $bitstream->getAssetstore();
     parent::delete($bitstream);
-    if($assetstore->getType() != MIDAS_ASSETSTORE_REMOTE && $this->getByChecksum($checksum) == false)
+    if(file_exists($path) && $assetstore->getType() != MIDAS_ASSETSTORE_REMOTE
+       && $this->getByChecksum($checksum) == false)
       {
       unlink($path);
       }
